first refactor for modular resources
