+++ ACE_PRIME/HelloAF/src/ace/framework/enums/message_type_enum.py
@@ -1,5 +1,6 @@
 from enum import Enum
 
+
 class MessageType(Enum):
     DATA = "DATA"
     PASS_THROUGH = "PASS_THROUGH"
