+++ ACE_PRIME/HelloAF/src/ace/framework/enums/operation_classification_enum.py
@@ -1,5 +1,6 @@
 from enum import Enum
 
+
 class OperationClassification(Enum):
     CREATE_REQUEST = "CREATE_REQUEST"
     ADD_TO_CONTEXT = "ADD_TO_CONTEXT"
