+++ ACE_PRIME/HelloAF/src/ace/resources/core/hello_layers/prompts/identities/l6_identity.md
