+++ ACE_PRIME/HelloAF/src/ace/resources/core/hello_layers/prompts/outputs/l6_south.md
