+++ ACE_PRIME/HelloAF/src/ace/resources/core/hello_layers/prompts/templates/l1_layer_instructions.md
