+++ ACE_PRIME/HelloAF/src/ace/resources/core/hello_layers/prompts/templates/l1_starting_instructions.md
