+++ ACE_PRIME/HelloAF/src/ace/resources/core/hello_layers/prompts/templates/layer_instructions.md
