+++ ACE_PRIME/HelloAF/src/ace/resources/core/hello_layers/prompts/templates/layer_log_message.md
