+++ ACE_PRIME/HelloAF/src/ace/resources/core/hello_layers/prompts/templates/op_log_message.md
